Update README.md
This a simplistic system for evaluating bid requests against advertising campaigns. Bid requests are messages that contain the following information:
• request ID
• page URL (e.g., http://apple.com/ca/store?item=1290)
• country of origin (e.g., CA, IT)
• and placement dimensions (e.g., 300x250)

Advertising campaigns are entities that are comprised of the following fields:
• campaign ID
• targeted country (e.g., CA, IT)
• targeted domain (e.g., apple.com)
• list of available ad dimensions (e.g., 300x250, 600x200)

There are 3 main objects in the system; Request --> BidRequest, Campaign --> AdCampaign and Evaluate. There are some other objects as well that help generate random ad campaigns and bid requests to run a simulation. The static and logic/operation part of the code have been separated into 2 different files.

The final output of the system is a JSON object that contains the following:
• defined campaigns
• list of evaluation results (include bid request and eligible campaign's ID)
• total bid requests processed
• and total evaluation time